ui: extract Team primitives (CopyableField, RedactedField, QueueTile)
