+++ packages/myco/ui/src/components/team/CopyableField.tsx
@@ -1,0 +1,32 @@
+import { useState, useCallback } from 'react';
+import { Copy, Check } from 'lucide-react';
+
+export function CopyableField({ label, value, mono = true }: { label: string; value: string; mono?: boolean }) {
+  const [copied, setCopied] = useState(false);
+
+  const handleCopy = useCallback(() => {
+    navigator.clipboard.writeText(value).then(() => {
+      setCopied(true);
+      setTimeout(() => setCopied(false), 2000);
+    });
+  }, [value]);
+
+  return (
+    <div className="space-y-1">
+      <span className="text-xs text-on-surface-variant">{label}</span>
+      <div className="flex items-center gap-2 group">
+        <span className={`text-sm text-on-surface break-all ${mono ? 'font-mono' : ''}`}>
+          {value}
+        </span>
+        <button
+          type="button"
+          onClick={handleCopy}
+          className="shrink-0 p-1 rounded text-on-surface-variant hover:text-on-surface opacity-0 group-hover:opacity-100 transition-opacity"
+          title="Copy to clipboard"
+        >
+          {copied ? <Check className="h-3.5 w-3.5 text-primary" /> : <Copy className="h-3.5 w-3.5" />}
+        </button>
+      </div>
+    </div>
+  );
+}
